requirements.txt
altair==6.0.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.4
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
googleapis-common-protos==1.72.0
greenlet==3.3.1
httplib2==0.31.2
idna==3.11
isodate
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.16.0
numpy==2.4.2
packaging==26.0
pandas==2.3.3
pillow==12.1.0
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
PyMySQL==1.1.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.46
streamlit==1.31.0
tenacity==9.1.3
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0